Add e2e tests for start up
