+++ package.json
@@ -28,6 +28,7 @@
   },
   "devDependencies": {
     "chai": "^3.5.0",
+    "chai-as-promised": "^7.1.1",
     "dotenv": "^2.0.0",
     "lodash": "^4.17.11",
     "mocha": "^2.5.3",
